feat: 채팅방 리스트 조회를 위한 새로운 Response \n RoomRes객체 구현

+++ backend/src/main/java/com/ssafy/db/entity/Room.java
@@ -27,7 +27,7 @@ public class Room {
     //시간 안해도 되나!?
 
     @Column(name="room_host_id", nullable = false)
-    private int roomHostId; //fk
+    private Long roomHostId; //fk
 
     @Column(name="room_title")
     private String roomTitle;
